Patch dev branch scan user storage

+++ src/screens/DashboardContainer/Files/Uploads/Audio/Audio.tsx
@@ -27,22 +27,32 @@ const Audio = ({navigation}: any) => {
   const isFocused = useIsFocused();
 
   useEffect(() => {
-    useGetUploadData('audio').then(fetchedData => {
-      setData(fetchedData as any[]);
-      if (fetchedData.length === 0) {
-        return Toast.show({
-          type: 'info',
-          text1: 'no data found',
+    if (isFocused) {
+      useGetUploadData('audio')
+        .then(fetchedData => {
+          setData(fetchedData as any[]);
+          if (fetchedData.length === 0) {
+            return Toast.show({
+              type: 'info',
+              text1: 'no data found',
+            });
+          } else {
+            return Toast.show({
+              text1: 'data fetched successfully',
+            });
+          }
+        })
+        .catch(e => {
+          console.log({error: e});
+          Toast.show({
+            type: 'error',
+            text1: 'cannot fetch files',
+          });
         });
-      } else {
-        return Toast.show({
-          text1: 'data fetched successfully',
-        });
-      }
-    });
+    }
 
     return () => {
-      if (removeFilesAfterFinish.length !== 0) {
+      if (removeFilesAfterFinish.length !== 0 && !isFocused) {
         for (const file of removeFilesAfterFinish) {
           try {
             RNFS.unlink(file).then(() => {
